still cleaning, create token normed, still got stuff to norm in Tokenizer folder
